moving public classes to package-private and removing module-info.java
