--- src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module roberta.tokenizer {
-    requires lombok;
-    requires com.google.common;
-    requires com.google.gson;
-    exports com.genesys.roberta.tokenizer;
-}
