+++ src/test/java/com/genesys/roberta/tokenizer/RobertaTokenizerResourcesTest.java
@@ -1,4 +1,4 @@
-package com.genesys.roberta.tokenizer.unit;
+package com.genesys.roberta.tokenizer;
 
 import com.genesys.roberta.tokenizer.BiGram;
 import com.genesys.roberta.tokenizer.RobertaTokenizer;
